Move inference to repository crate
