+++ crates/turborepo-repository/src/lib.rs
@@ -1,4 +1,5 @@
 #![feature(error_generic_member_access)]
 
+pub mod inference;
 pub mod package_json;
 pub mod package_manager;
